Fix project build
-Removed some arc issues for de alloc, release as arc automatically
handles this
-fixed project structure to run application

+++ EMBRACE_Client/ITS/BayesKnowledgeTracer/ManipulationAnalyser.m
@@ -809,7 +809,7 @@
         feedback = @"Read";
     }
     
-    NSLog(@" Error type = %@ || Error count = %d || Type of feedback = %@ ", errorType, errorCount, feedback);
+    NSLog(@" Error type = %@ || Error count = %ld || Type of feedback = %@ ", errorType, (long)errorCount, feedback);
     self.currentFeedback = feedbackObjc;
     
 }
